feat(plugin): adapter plugin -- greptimedb and questdb
