+++ configure/etc/conf/plugins.properties
@@ -41,6 +41,8 @@ datacap-plugin-trino=plugin/datacap-plugin-trino/pom.xml
 datacap-plugin-doris=plugin/datacap-plugin-doris/pom.xml
 datacap-plugin-starrocks=plugin/datacap-plugin-starrocks/pom.xml
 datacap-plugin-hologres=plugin/datacap-plugin-hologres/pom.xml
+datacap-plugin-greptimedb=plugin/datacap-plugin-greptimedb/pom.xml
+datacap-plugin-questdb=plugin/datacap-plugin-questdb/pom.xml
 datacap-convert-txt=convert/datacap-convert-txt/pom.xml
 datacap-convert-json=convert/datacap-convert-json/pom.xml
 datacap-convert-xml=convert/datacap-convert-xml/pom.xml
