--- plugin/datacap-http-greptime/src/main/java/io/edurt/datacap/plugin/http/greptime/response/GreptimeDBResponse.java
@@ -1,24 +1,0 @@
-package io.edurt.datacap.plugin.http.greptime.response;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
-import io.edurt.datacap.plugin.http.greptime.response.record.Output;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.ToString;
-
-import java.util.List;
-
-@Data
-@ToString
-@NoArgsConstructor
-@AllArgsConstructor
-@JsonIgnoreProperties(ignoreUnknown = true)
-@SuppressFBWarnings(value = {"EI_EXPOSE_REP", "EI_EXPOSE_REP2"})
-public class GreptimeDBResponse
-{
-    private int code;
-    private String error;
-    private List<Output> output;
-}
