--- plugin/datacap-http-greptime/src/main/java/io/edurt/datacap/plugin/http/greptime/response/record/ColumnSchema.java
@@ -1,20 +1,0 @@
-package io.edurt.datacap.plugin.http.greptime.response.record;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.ToString;
-
-@Data
-@ToString
-@NoArgsConstructor
-@AllArgsConstructor
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class ColumnSchema
-{
-    private String name;
-    @JsonProperty(value = "data_type")
-    private String dataType;
-}
