--- plugin/datacap-http-greptime/src/main/java/io/edurt/datacap/plugin/http/greptime/response/record/Output.java
@@ -1,19 +1,0 @@
-package io.edurt.datacap.plugin.http.greptime.response.record;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.ToString;
-
-@Data
-@ToString
-@NoArgsConstructor
-@AllArgsConstructor
-@JsonIgnoreProperties(ignoreUnknown = true)
-@SuppressFBWarnings(value = {"PA_PUBLIC_PRIMITIVE_ATTRIBUTE", "EI_EXPOSE_REP2"})
-public class Output
-{
-    public Records records;
-}
