--- plugin/datacap-http-questdb/src/main/java/io/edurt/datacap/plugin/http/questdb/response/Schema.java
@@ -1,18 +1,0 @@
-package io.edurt.datacap.plugin.http.questdb.response;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.ToString;
-
-@Data
-@ToString
-@NoArgsConstructor
-@AllArgsConstructor
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class Schema
-{
-    private String name;
-    private String type;
-}
