+++ plugin/datacap-plugin-greptimedb/src/main/java/io/edurt/datacap/plugin/GreptimeDBPlugin.java
@@ -1,0 +1,5 @@
+package io.edurt.datacap.plugin;
+
+public class GreptimeDBPlugin
+        extends Plugin
+{}
