+++ plugin/datacap-plugin-questdb/src/main/java/io/edurt/datacap/plugin/QuestDBPlugin.java
@@ -1,0 +1,5 @@
+package io.edurt.datacap.plugin;
+
+public class QuestDBPlugin
+        extends Plugin
+{}
