+++ plugin/datacap-plugin-questdb/src/main/java/io/edurt/datacap/plugin/response/Schema.java
@@ -1,0 +1,18 @@
+package io.edurt.datacap.plugin.response;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
+
+@Data
+@ToString
+@NoArgsConstructor
+@AllArgsConstructor
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class Schema
+{
+    private String name;
+    private String type;
+}
